[REFACTOR] 팀원신청 승인 전에 취소 후 예외 출력시, applicantRepository에서 applicant를 받아오도록 수정
